remove popover trigger. make the target a property of the popover (element id)
