+++ src/index.js
@@ -4,6 +4,7 @@ import Popover from './Popover';
 import PopoverBody from './PopoverBody';
 import PopoverFooter from './PopoverFooter';
 import ProcessingIndicator from './ProcessingIndicator';
+import PopoverOverlay from './PopoverOverlay';
 import ProgressBar from './ProgressBar';
 import StatusIndicator from './StatusIndicator';
 
@@ -13,6 +14,7 @@ module.exports = {
   Popover,
   PopoverBody,
   PopoverFooter,
+  PopoverOverlay,
   ProcessingIndicator,
   ProgressBar,
   StatusIndicator
